Refactor project structure into application and CLI
