+++ src/Monoalphabetische.Application/Decrypt.cs
@@ -1,44 +1,44 @@
 ﻿using System.Text;
 
-namespace Monoalphabetische.Substitution;
+namespace Monoalphabetische.Application;
 
-public static class Encrypt
-{
-    public static void _Encrypt(Message message)
+public static class Decrypt
+{ 
+    public static void _Decrypt(Message message)
     {
         if (!message.IsValid)
         {
             throw new ArgumentException("Message is invalid. Maybe the encrypted/decrypted message contains an unsupported character or an unsupported key.");
         }
 
-        if(string.IsNullOrWhiteSpace(message.DecryptedMessagae))
+        if (string.IsNullOrWhiteSpace(message.EncryptedMessage))
         {
             throw new ArgumentException("Can't encrypt message because there is no decrypted message defined.");
         }
 
-        if(message.Key == null)
+        if (message.Key == null)
         {
             throw new ArgumentException("Can't encrypt message because there is not key given.");
         }
 
         StringBuilder stringBuilder = new StringBuilder();
-        foreach (char character in message.DecryptedMessagae) 
+        foreach (char character in message.EncryptedMessage)
         {
             int index = getNewAlphabethIndex(character, Convert.ToInt32(message.Key));
             stringBuilder.Append(MessageHelper.Alphabeth[index]);
         }
 
-        message.EncryptedMessage = stringBuilder.ToString();
+        message.DecryptedMessagae = stringBuilder.ToString();
     }
 
     private static int getNewAlphabethIndex(char character, int key)
     {
         int baseIndex = Array.IndexOf(MessageHelper.Alphabeth, character);
-        int newIndex = baseIndex + key;
+        int newIndex = baseIndex - key;
 
-        if (newIndex >= MessageHelper.Alphabeth.Length)
+        if (newIndex < 0)
         {
-            newIndex -= MessageHelper.Alphabeth.Length;
+            newIndex = MessageHelper.Alphabeth.Length + newIndex;
         }
 
         return newIndex;
