+++ src/Monoalphabetische.Application/Encrypt.cs
@@ -1,44 +1,44 @@
 ﻿using System.Text;
 
-namespace Monoalphabetische.Substitution;
+namespace Monoalphabetische.Application;
 
-public static class Decrypt
-{ 
-    public static void _Decrypt(Message message)
+public static class Encrypt
+{
+    public static void _Encrypt(Message message)
     {
         if (!message.IsValid)
         {
             throw new ArgumentException("Message is invalid. Maybe the encrypted/decrypted message contains an unsupported character or an unsupported key.");
         }
 
-        if (string.IsNullOrWhiteSpace(message.EncryptedMessage))
+        if(string.IsNullOrWhiteSpace(message.DecryptedMessagae))
         {
             throw new ArgumentException("Can't encrypt message because there is no decrypted message defined.");
         }
 
-        if (message.Key == null)
+        if(message.Key == null)
         {
             throw new ArgumentException("Can't encrypt message because there is not key given.");
         }
 
         StringBuilder stringBuilder = new StringBuilder();
-        foreach (char character in message.EncryptedMessage)
+        foreach (char character in message.DecryptedMessagae) 
         {
             int index = getNewAlphabethIndex(character, Convert.ToInt32(message.Key));
             stringBuilder.Append(MessageHelper.Alphabeth[index]);
         }
 
-        message.DecryptedMessagae = stringBuilder.ToString();
+        message.EncryptedMessage = stringBuilder.ToString();
     }
 
     private static int getNewAlphabethIndex(char character, int key)
     {
         int baseIndex = Array.IndexOf(MessageHelper.Alphabeth, character);
-        int newIndex = baseIndex - key;
+        int newIndex = baseIndex + key;
 
-        if (newIndex < 0)
+        if (newIndex >= MessageHelper.Alphabeth.Length)
         {
-            newIndex = MessageHelper.Alphabeth.Length + newIndex;
+            newIndex -= MessageHelper.Alphabeth.Length;
         }
 
         return newIndex;
