+++ src/Monoalphabetische.Application/Message.cs
@@ -1,4 +1,4 @@
-﻿namespace Monoalphabetische.Substitution;
+﻿namespace Monoalphabetische.Application;
 
 public class Message
 {
