+++ src/Monoalphabetische.Cli/Monoalphabetische.Cli.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net9.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Monoalphabetische.Application\Monoalphabetische.Application.csproj" />
+  </ItemGroup>
+
+</Project>
