--- src/Monoalphabetische.Substitution.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Monoalphabetische.Substitution", "Monoalphabetische.Substitution\Monoalphabetische.Substitution.csproj", "{2ADC3125-047E-4530-91AE-803B27E73E1E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2ADC3125-047E-4530-91AE-803B27E73E1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2ADC3125-047E-4530-91AE-803B27E73E1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2ADC3125-047E-4530-91AE-803B27E73E1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2ADC3125-047E-4530-91AE-803B27E73E1E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
