+++ src/Monoalphabetische.sln
@@ -1,0 +1,27 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Monoalphabetische.Cli", "Monoalphabetische.Cli\Monoalphabetische.Cli.csproj", "{FC784AC1-018F-4BFF-8016-B0DE810609F1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Monoalphabetische.Application", "Monoalphabetische.Application\Monoalphabetische.Application.csproj", "{D70BA1CF-9383-4310-B63C-769EC1211B89}"
+EndProject
+Global
+GlobalSection(SolutionConfigurationPlatforms) = preSolution
+Debug|Any CPU = Debug|Any CPU
+Release|Any CPU = Release|Any CPU
+EndGlobalSection
+GlobalSection(ProjectConfigurationPlatforms) = postSolution
+{FC784AC1-018F-4BFF-8016-B0DE810609F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+{FC784AC1-018F-4BFF-8016-B0DE810609F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+{FC784AC1-018F-4BFF-8016-B0DE810609F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+{FC784AC1-018F-4BFF-8016-B0DE810609F1}.Release|Any CPU.Build.0 = Release|Any CPU
+{D70BA1CF-9383-4310-B63C-769EC1211B89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+{D70BA1CF-9383-4310-B63C-769EC1211B89}.Debug|Any CPU.Build.0 = Debug|Any CPU
+{D70BA1CF-9383-4310-B63C-769EC1211B89}.Release|Any CPU.ActiveCfg = Release|Any CPU
+{D70BA1CF-9383-4310-B63C-769EC1211B89}.Release|Any CPU.Build.0 = Release|Any CPU
+EndGlobalSection
+GlobalSection(SolutionProperties) = preSolution
+HideSolutionNode = FALSE
+EndGlobalSection
+EndGlobal
